+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.7.0] - Unreleased
 - Added option to not show shell commands in the interactive view.
+- Added option to search only labels.
 
 ## [0.6.0] - 2021-02-09
 - Added bold label display option.